+++ Changelog.md
@@ -7,6 +7,10 @@
 - Only offer a Destiny Card for a **Defend Approved Action** if the target is doing an "Active Defense" this round.
 - Slight updates to Possibility Selection Dialog.
 
+### Bug Fixes
+
+- Stop the tabs of the Actor sheet flickering on each update.
+
 ## 13.8.2 - Fixes for Perks, Cosms and Secrets
 
 ### Improvements